requirements.txt
pyTelegramBotAPI
python-dotenv
gspread
google-auth
PyYAML
pytest
